pyproject.toml
[project]
name = "structlog-extras"
description = "Various structlog additions, mainly focused on better stdlib integration"
version = "0.4.0"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Alexey Shokov" }]
keywords = ["logging", "structlog"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Typing :: Typed",
    "Topic :: Software Development",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
]
dependencies = ["structlog >=24.0"]

# See https://docs.pypi.org/project_metadata/#general-url
# See also https://daniel.feldroy.com/posts/2023-08-pypi-project-urls-cheatsheet
[project.urls]
Changelog = 'https://github.com/alexeyshockov/structlog-extras/blob/main/CHANGELOG.md'

[dependency-groups]
dev = ["python-dotenv", "rich", "orjson", "opentelemetry-exporter-otlp"]

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"

[tool.uv]
build-backend.module-root = "" # Flat layout
build-backend.module-name = "structlog_extras"

[tool.ruff]
line-length = 120
format.docstring-code-format = true

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
    "Q", # flake8-quotes
]
[tool.ruff.lint.pydocstyle]
convention = "google"
